Custom pandoc script to convert .md to .tex
In this commit, I create a custom pandoc writer (i.e. conversion
ruleset) in order to create a conversion script that converts markdown
(specifically, pandoc-style markdown with YAML metadata) into my
idiosyncratic LaTeX format.

This is done with a view towards eventually writing all journal entries
entirely in markdown, which can be converted losslessly through the
custom pandoc writer into LaTeX that is then compiled into PDF.

Because I am using the new-style pandoc writer syntax, in order for the
script to work a pandoc version of 2.17.2 or greater must be installed.
For comparison, as of 2022-11-11 the latest pandoc release is version
2.19.2.