Make dropcaps.py filter more robust w/ validation
By adding additional checks to makesure that the first element of the
first paragraph in the pandoc AST is of a type Str (and not anything
else, such as a RawInline (i.e. LaTeX command) or Note (i.e. citation)),
we can make sure that the dropcaps filter is much more robust to these
unusual edge cases.